+++ internal/parsergen/backend/python/doc.go
@@ -1,0 +1,3 @@
+// Package python provides a parser generator backend which takes a parse table as input and outputs Python source
+// code.
+package python