+++ internal/scannergen/backend/python/doc.go
@@ -1,0 +1,2 @@
+// Package python provides a scanner generator backend which takes a DFA as input and outputs Python source code.
+package python